requirements.txt
# pip install -r requirements.txt
# pip freeze > requirements.txt
APScheduler==3.11.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Deprecated==1.1.5
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyGithub==2.5.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
